+++ Smokebot/run_smokebot.sh
@@ -290,11 +290,15 @@ COMPILER="-I $COMPILER"
 if [ "$KILL_SMOKEBOT" == "1" ]; then
   if [ -e $smokebot_pid ]; then
     PID=`head -1 $smokebot_pid`
-    JOBS=$(LIST_DESCENDANTS $PID)
-    echo killing processes invoked by smokebot: $JOBS
-    kill -9 $JOBS
     echo "killing smokebot (PID=$PID)"
     kill -9 $PID
+
+    JOBS=$(LIST_DESCENDANTS $PID)
+    if [ "$JOBS" != "" ]; then
+      echo killing processes invoked by smokebot: $JOBS
+      kill -9 $JOBS
+    fi
+
     if [ "$QUEUE" == "none" ]; then
       cd $CURDIR/../Scripts
       ./killppids.sh ../Smokebot/scriptfiles
@@ -306,7 +310,9 @@ if [ "$KILL_SMOKEBOT" == "1" ]; then
         qdel $JOBIDS
       fi
     fi
+
     echo smokebot process $PID killed
+    rm -f $smokebot_pid
   else
     echo smokebot not running
   fi
